third_party/WebKit/LayoutTests/virtual/stable/webexposed/web-animations-api-expected.txt

Check that Web Animations API methods are not available without experimental web platform features enabled. Expected to fail in LayoutTests/webexposed and pass in LayoutTests/virtual/stable/webexposed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Element.animate() is not exposed.
PASS Animation constructor is not exposed.
PASS successfullyParsed is true

TEST COMPLETE
